requirements.txt
flask==3.0.0
flask-cors==4.0.0
mediapipe==0.10.9
opencv-python-headless==4.9.0.80
Pillow==10.2.0
numpy==1.26.4
gunicorn==21.2.0
